requirements.txt
alembic==1.5.8
aniso8601==9.0.1
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
coverage==5.5
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
# Flask-JWT-Extended==4.1.0
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
six==1.15.0
SQLAlchemy==1.4.4
Werkzeug==1.0.1
